Переделан сервак, теперь архитектура разделена по сущностям. Отдельно контроллеры и репозитории.
