+++ server/src/initDb.js
@@ -1,6 +1,7 @@
 // Модуль для работы с БД
-const { DB } = require("./db");
-const db = new DB("../test/cadets.db");
+const { db } = require("./db");
+const cadetRepo = new require("./repositories/cadetRepo")(db);
+
 
 const {
     getCourseList
@@ -9,9 +10,6 @@ const {
 // Асинхронная функция для обработки
 async function processData() {
     try {
-        // Подключение к базе данных
-        await db.connect();
-        await db.createTables();
         // Получаем данные о курсах
         const сourses = [getCourseList('../test/1курс/группы/'), getCourseList('../test/2курс/группы/'),
                         getCourseList('../test/3курс/группы/'), getCourseList('../test/4курс/группы/'), getCourseList('../test/5курс/группы/')];
@@ -22,7 +20,7 @@ async function processData() {
         for (let course = 0; course < 5; course++) {
             for (let group in сourses[course]) {
                 for (let cadet of сourses[course][group]) {
-                    db.insertCadet(course + 1, group, serial, cadet['Воинское звание'], cadet['ФИО'],
+                    cadetRepo.insert(course + 1, group, serial, cadet['Воинское звание'], cadet['ФИО'],
                         cadet['Дата рождения'], cadet['Номер телефона'], cadet['Личный номер']);
                 }
             }
